split out data file view tests
